replace uin8_t by faster uint32_t

+++ lib/gcode/gcodemodalgroup.h
@@ -8,7 +8,7 @@
 #pragma once
 #include <stdint.h>
 
-enum class modalGroup : uint8_t
+enum class modalGroup : uint32_t
     {
     Group0,
     Motion,
@@ -31,69 +31,69 @@ enum class modalGroup : uint8_t
     };
 constexpr uint32_t nmbrModalGroups = static_cast<uint32_t>(modalGroup::nmbrModalGroups);
 
-enum class modalGroup0 : uint8_t
+enum class modalGroup0 : uint32_t
     {
     G4, G10, G28, G30, G53,
     //  G92, G92_1, G92_2, G92_3
     nmbrModes
     };
 
-enum class modalGroupMotion : uint8_t
+enum class modalGroupMotion : uint32_t
     {
     G0, G1, G2, G3,
     nmbrModes
     };
 
-enum class modalGroupFeedRate : uint8_t
+enum class modalGroupFeedRate : uint32_t
     {
     G94, G93,
     nmbrModes
     };
 
-enum class modalGroupSpindle : uint8_t
+enum class modalGroupSpindle : uint32_t
     {
     M5, M3, M4,
     nmbrModes
     };
 
-enum class modalGroupUnit : uint8_t
+enum class modalGroupUnit : uint32_t
     {
     G21, G20,
     nmbrModes
     };
 
-enum class modalGroupPlane : uint8_t
+enum class modalGroupPlane : uint32_t
     {
     G17, G18, G19,
     nmbrModes
     };
 
-enum class modalGroupDistance : uint8_t
+enum class modalGroupDistance : uint32_t
     {
     G90, G91,
     nmbrModes
     };
 
-enum class modalGroupCoordinateSet : uint8_t
+enum class modalGroupCoordinateSet : uint32_t
     {
     G53, G54, G55, G56, G57, G58, G59, G59_1, G59_2, G59_3,
     nmbrCoordinateSets
     };
 constexpr uint32_t nmbrCoordinateSets = static_cast<uint32_t>(modalGroupCoordinateSet::nmbrCoordinateSets);
 
-enum class modalGroupOverrides : uint8_t
+enum class modalGroupOverrides : uint32_t
     {
     M48, M49,
     nmbrModes
     };
 
-enum class modalGroupCoolantFlood : uint8_t
+enum class modalGroupCoolantFlood : uint32_t
     {
     M9, M7,
     nmbrModes
     };
 
-enum class modalGroupCoolantMist : uint8_t
+enum class modalGroupCoolantMist : uint32_t
     {
     M9, M8,
     nmbrModes
